// sdk/plugin-protocol/types/PluginInspection.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PluginArtifactSummary } from "./PluginArtifactSummary.js";
import type { PluginCatalogItem } from "./PluginCatalogItem.js";
import type { PluginManifest } from "./PluginManifest.js";
import type { RevisionId } from "./RevisionId.js";

export type PluginInspection = { summary: PluginCatalogItem, manifest: PluginManifest, parent: RevisionId | null, source_file_count: number, artifacts: Array<PluginArtifactSummary>, };